Add files via upload
This is the milestone 2 project. I have created a class for inventory items and have used the existing shell as a driver. The user can click the add button to add an inventory item object defined in the class to a list<> which stores the objects for now. A messagebox will display the item added by accessing the list<> by properties defined in the InventoryItem Class.